pyproject.toml
[project]
name = "libapi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.5",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "drf-yasg>=1.21.10",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
]
